feature: Add type_decorators module
The database package encapsulates the basic logic of working with
databases and contains subpackages that implement more specific logic
related to libraries of the same name, such as sqlalchemy.

The type_decorator module is based on part of the code of the
old models module, and completely transfers the IntEnum and JSONArray
classes from there for further use in type_annotation_map.